+++ ebi.h
@@ -1,16 +1,14 @@
-//Big number class definition. This is just an example.
-
 /**************************************************
 By vegetablebird 2018.03.09
   	Big-endian is implemented
 	ebi(0) has positive sign and 1 digit
 	int type is used for indexing, thus MAX_NUM_OF_BITS should not larger than INT_MAX
-	I assume nobody will explicitly call the deconstructure
+	I assume nobody will explicitly call the deconstructor
 		i.e. should not exists a instance with NULL data array
 **************************************************/
 
-#ifndef __BIGNUMBER__
-#define __BIGNUMBER__
+#ifndef __EXTREMELYEASYTOUSEBIGINTEGER__
+#define __EXTREMELYEASYTOUSEBIGINTEGER__
 
 #include <iostream>
 //#include <cstdint>
@@ -22,47 +20,40 @@ typedef unsigned char uint8_t;
 
 class ebi {
 	private:
-		bool sgn; // sign
-		unsigned int num_of_bits; // N_xdigits
+		enum {negative, positive} sign;
+		unsigned int N_xdigits;
 		uint8_t *data;
 
-		const static bool positive = true;
-		const static bool negative = false;
-
+		inline void base_initialization();
 		inline ebi base_addition(const ebi &a, const ebi &b) const;
 		inline ebi base_subtraction(const ebi& a, const ebi& b) const;
 		inline bool base_lessthan(const ebi& a, const ebi& b) const;
 
-		inline bool isPositive() const;
-		inline bool isNegative() const;
-
 	public:
-		//constructors
+		//constructors and deconstructor
 		ebi();
 		ebi(int);  //directly convert from an int
-		ebi(bool, unsigned int, uint8_t*);
+		ebi(bool isPositive, unsigned int nXDigits, uint8_t* rawData);
 		ebi(const ebi&);
 		ebi(const char*);
-
-		//deconstructor
 		~ebi();
 
-		//overloaded arithmetic operators as member functions
-		ebi operator+(ebi);
-		ebi operator-(ebi);
-		ebi operator*(ebi);
-		ebi operator/(ebi); //integer division: 3/2==1
-		ebi operator%(ebi);
+		//arithmetic operators
+		ebi operator+(const ebi&) const;
+		ebi operator-(const ebi&) const;
+		ebi operator*(const ebi&) const;
+		ebi operator/(const ebi&) const; //integer division: 3/2==1
+		ebi operator%(const ebi&) const; //by C99, a == (a/b*b) + a%b
 		ebi operator-() const;
 		ebi operator<<(unsigned int) const;
 		ebi operator>>(unsigned int) const;
 
 		//interface functions
-		void Print();
-		void GetData(bool&, unsigned int&, uint8_t*);
-		bool get_sgn() const;
-		unsigned int get_num_of_bits() const;
-		uint8_t get_data(unsigned int) const;
+		void Print(); // discarded
+		void GetData(bool&, unsigned int&, uint8_t*); // discarded
+		bool get_sign() const;
+		unsigned int get_N_xdigits() const;
+		uint8_t get_data(unsigned int) const; // get_xdigit, operator[]
 		friend ostream& operator<<(ostream&, const ebi&);
 		friend istream& operator>>(istream&, ebi&);
 
@@ -97,13 +88,13 @@ ebi rand(const unsigned digits);
 ebi abs(const ebi &);
 ebi pow(ebi base, unsigned exponent);
 
-ebi operator+(int n, ebi bn);
-ebi operator-(int n, ebi bn);
-ebi operator*(int n, ebi bn);
-ebi operator/(int n, ebi bn);
-ebi operator%(int n, ebi bn);
-bool operator<(int n, ebi bn);
-bool operator!=(int n, ebi bn);
-bool operator==(int n, ebi bn);
+ebi operator+(int n, const ebi &bn);
+ebi operator-(int n, const ebi &bn);
+ebi operator*(int n, const ebi &bn);
+ebi operator/(int n, const ebi &bn);
+ebi operator%(int n, const ebi &bn);
+bool operator<(int n, const ebi &bn);
+bool operator!=(int n, const ebi &bn);
+bool operator==(int n, const ebi &bn);
 
 #endif