+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # Changelog
 
 ## [Unreleased]
-- Add demo code to README.md
+- Add Demo and Features in README.md
+- Modify Operator%() by C99 definition, a == (a/b*b) + a%b
+- Rename num_of_bits to N_xdigits
+- Rename __BIGNUMBER__ to __EXTREMELYEASYTOUSEBIGINTEGER__
+- Remove DEBUG_BIGNUMBER
+- Support decimal and hexadecimal output
+- Remove isPositive() and isNegative()
+- Rename bool sgn to enum sign
+- Remove constants positive and negative
+- Modify INITIALIZE() to base_initialization()